Add flag to skip app deletion after tests

+++ app_test_suite/steps/scenarios/simple.py
@@ -230,6 +230,8 @@ class SimpleTestScenario(BuildStep, ABC):
         finally:
             if not get_config_value_by_cmd_line_option(
                 config, BaseTestScenariosFilteringPipeline.KEY_CONFIG_OPTION_SKIP_DEPLOY_APP
+            ) and not get_config_value_by_cmd_line_option(
+                config, BaseTestScenariosFilteringPipeline.KEY_CONFIG_OPTION_SKIP_DELETE_APP
             ):
                 self._delete_app(config, context)
